Adds Experience section to HomePage and updates user info
Integrates an Experience component into the HomePage to showcase professional experience.

Updates ExperienceInfo to include a new entry for a Frontend Development Intern position, while commenting out the previous System Engineer role to streamline the focus on current experiences.

This change enhances the overall presentation of the user's skills and provides a clearer depiction of their career progression.